+++ src/accessory/climate/heaterCooler.ts
@@ -57,12 +57,7 @@ export class HeaterCoolerAccessory extends ActiveClimateAccessory<HeaterCoolerCo
       'topicSetTargetHeaterCoolerState', this.onSetTargetState.bind(this))
       ?.setProps({ validValues: validTargetStates.map((key) => this.TARGET_STATE_MAP.get(key)!) });
 
-    this.addTopicHandlers(
-      FilterMaintenance.topicHandlers(
-        Service, Characteristic, log, config.disableLogging, accessory,
-        this.accessoryService, config, this.name, this.publish.bind(this),
-      ),
-    );
+    this.addTopicHandlers(FilterMaintenance.topicHandlers(Service, this, config));
   }
 
   protected getAccessoryType(): AccessoryType {